requirements.txt
aiogram==2.21
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
altgraph==0.17.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.4
distlib==0.3.5
Django==4.0.6
filelock==3.7.1
frozenlist==1.3.1
future==0.18.2
idna==3.3
multidict==6.0.2
pefile==2021.9.3
platformdirs==2.5.2
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
python-decouple==3.6
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.15.1
yarl==1.8.1
youtube-search==2.1.1